# api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: Friend Tracker API
  description: API for tracking friends' locations and sending invitations
  version: "1.0.0"
host: example.com
basePath: /api
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /auth:
    post:
      summary: Authenticate user via Google
      description: Authenticate the user using their Google account and return an access token
      operationId: authenticateUser # Cette ligne lie la route à la fonction dans api/controllers/auth.js
      parameters:
        - name: id_token
          in: body
          description: The ID token obtained from Google after successful authentication
          required: true
          schema:
            type: object
            properties:
              id_token:
                type: string
      responses:
        200:
          description: Access token for the user
          schema:
            type: object
            properties:
              access_token:
                type: string
        400:
          description: Invalid ID token
        500:
          description: Internal server error
  /users:
    put:
      summary: Update user's avatar and name
      description: Update the user's avatar and name with the specified values
      operationId: updateUser  # Cette ligne lie la route à la fonction dans api/controllers/users.js
      parameters:
        - name: authorization
          in: header
          description: Access token to authorize request
          required: true
          type: string
        - name: avatar
          in: body
          required: false
          schema:
            type: object
            properties:
              avatar:
                type: string
                description: The new avatar image for the user
              name:
                description: The new name for the user
                type: string
      responses:
        200:
          description: User info
          schema:
            type: object
            properties:
              id:
                type: integer
                description: The ID of the user
              name:
                type: string
                description: The name of the user
              avatar:
                type: string
                description: The URL of the user's avatar image
              email:
                type: string
                description: The email address of the user
        401:
          description: Unauthorized request
        500:
          description: Internal server error
...